refactor(torghut): extract receipt normalization primitives
